Revert "Continuous ca"
