--- CAs/Continuous/life.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="utf-8" />
-  <title>WebGPU Life</title>
-  <link rel="stylesheet" href="style.css"> <!-- Link to the CSS file -->
-</head>
-
-<div class="text-column">
-  <h1>Game of Life</h1>
-</div>
-
-<body>
-  <div class="centre">
-    <!-- Buttons corresponding to key clicks -->
-    <button id="play">Play/Pause</button>
-    <button id="next">Next Frame</button>
-
-    <div class="grid">
-      <input type="number" id="kernel1" placeholder="0.11">
-      <input type="number" id="kernel2" placeholder="0.11">
-      <input type="number" id="kernel3" placeholder="0.11">
-      <input type="number" id="kernel4" placeholder="0.11">
-      <input type="number" id="kernel5" placeholder="0.11">
-      <input type="number" id="kernel6" placeholder="0.11">
-      <input type="number" id="kernel7" placeholder="0.11"> 
-      <input type="number" id="kernel8" placeholder="0.11">
-      <input type="number" id="kernel9" placeholder="0.11">
-      <button id="submitKernel">Submit Kernel</button>
-    </div>
-    
-    <!-- Input Box for Speed -->
-    <input type="text" id="speedInputBox" placeholder="1-100:3">
-    <button id="speedInput">Submit Speed </button>
-  </div>
-
-  <!-- Canvas containing cellular automata -->
-  <div> <canvas id="canvas" width="1024" height="1024" /> </div>
-  <script type="module" src="life.js"></script>
-</body>
-
-</html>
