--- CAs/Larger/life.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="utf-8" />
-  <title>WebGPU Life</title>
-  <link rel="stylesheet" href="style.css"> <!-- Link to the CSS file -->
-</head>
-
-<div class="text-column">
-  <h1>Game of Life</h1>
-</div>
-
-<body>
-  <div class="centre">
-    <!-- Buttons corresponding to key clicks -->
-    <button id="play">Play/Pause</button>
-    <button id="next">Next Frame</button>
-
-    <!-- Input Box for Simulation Parameters -->
-    <!-- Thinking of converting to boxes as follows for user conveniance: 
-      R[<input>], C[<input>], S[<input>], B[<input>], N[<input>]
-      We can then restrict input to numerical only with html I believe.
-    -->
-    <input type="text" id="simulationInput" placeholder="Enter rulestring :3">
-    <button id="submitInput">Submit Rule</button>
-
-    <!-- Input Box for Speed -->
-    <input type="text" id="speedInputBox" placeholder="1-100:3">
-    <button id="speedInput">Submit Speed </button>
-  </div>
-
-  <!-- Canvas containing cellular automata -->
-  <div> <canvas id="canvas" width="1024" height="1024" /> </div>
-  <script type="module" src="life.js"></script>
-</body>
-
-</html>
