--- CAs/Larger/style.css
@@ -1,48 +1,0 @@
-/* General Body Styling */
-body {
-    font-family: Arial, sans-serif;
-    background-color: #f0f0f0;
-    margin: 0;
-    padding: 0;
-}
-
-/* Styling for the text column */
-.text-column {
-    text-align: center;
-    margin-top: 20px;
-}
-
-.text-column h1 {
-    color: #333;
-}
-
-/* Canvas Styling */
-canvas {
-    display: block;
-    margin: 20px auto;
-    border: 1px solid #ddd;
-}
-
-.centre {
-    margin: auto;
-    width: 50%;
-    padding: 10px;
-}
-
-/* Button Styling */
-button {
-    background-color: pink;
-    color: white;
-    border: none;
-    padding: 10px 20px;
-    margin: 10px;
-    cursor: pointer;
-    font-size: 16px;
-    border-radius: 5px;
-    transition: background-color 0.3s;
-}
-
-button:hover {
-    background-color: #ff69b4;
-    /* A darker shade of pink for hover effect */
-}
