+++ CAs/LifeLike/life.js
@@ -3,12 +3,8 @@ import { guiShader } from "./guiShader.js";
 import { computeShader } from "./computeShader.js";
 
 // import static manager classes
-import EventManager from "../Shared/managers/EventManager.js";
-import DeviceManager from "../Shared/managers/DeviceManager.js";
-
-import BufferManager from "../Shared/managers/BufferManager.js";
-import { parseRuleString } from "./Parse.js";
-import startingPatterns from "./startingPatterns.js";
+import EventManager from "./managers/EventManager.js";
+import DeviceManager from "./managers/DeviceManager.js";
 // import PipelineManager from "./managers/PipelineManager.js";
 // construct static classes lol
 await DeviceManager.staticConstructor();
@@ -16,31 +12,85 @@ const device = DeviceManager.device
 const canvas = DeviceManager.canvas
 
 // Set global variables
-const WORKGROUP_SIZE = 16; // only 1, 2, 4, 8, 16 work. higher is smoother. // There is a limitation though to some pcs/graphics cards
-const INITIAL_STATE = startingPatterns.cgl_B29;
-const GRID_SIZE = INITIAL_STATE.minGrid*2;//document.getElementById("canvas").getAttribute("width"); // from canvas size in life.html
-// [
-//     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-//     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-//     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-//     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-//     0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0,
-//     0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1,
-//     0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1,
-//     0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0,
-//     0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0,
-//     0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1,
-//     0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1,
-//     0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0,
-//     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-//     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-//     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-//     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-// ];
+const GRID_SIZE = document.getElementById("canvas").getAttribute("width"); // from canvas size in life.html
+const WORKGROUP_SIZE = 16; // only 1, 2, 4, 8, 16 work. higher is smoother.
+const INITIAL_STATE = [
+    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+    0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0,
+    0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1,
+    0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1,
+    0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0,
+    0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0,
+    0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1,
+    0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1,
+    0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0,
+    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+];
 
-EventManager.ruleString = INITIAL_STATE.rule;
+let step = 0
 
-const SQUARE_VERTICIES = new Float32Array([
+// DEVICE SETUP - ran into issues making it a func
+
+// Uniform grid
+const uniformArray = new Float32Array([GRID_SIZE, GRID_SIZE]);
+const uniformBuffer = device.createBuffer({
+    label: "Grid Uniforms",
+    size: uniformArray.byteLength,
+    usage: GPUBufferUsage.UNIFORM | GPUBufferUsage.COPY_DST,
+});
+
+device.queue.writeBuffer(uniformBuffer, 0, uniformArray);
+
+// Cell state arrays
+const cellStateArray = new Uint32Array(GRID_SIZE * GRID_SIZE);
+const cellStateStorage = [
+    device.createBuffer({
+        label: "Cell State A",
+        size: cellStateArray.byteLength,
+        usage: GPUBufferUsage.STORAGE | GPUBufferUsage.COPY_DST,
+    }),
+
+    device.createBuffer({
+        label: "Cell State B",
+        size: cellStateArray.byteLength,
+        usage: GPUBufferUsage.STORAGE | GPUBufferUsage.COPY_DST,
+    })
+];
+
+// write to buffer A
+for (let i = 0; i < cellStateArray.length; ++i) {
+    // cellStateArray[i] = Math.random() > 0.6 ? 1 : 0; // random starting position
+    cellStateArray[i] = INITIAL_STATE[i];
+}
+device.queue.writeBuffer(cellStateStorage[0], 0, cellStateArray);
+
+// write to buffer B
+for (let i = 0; i < cellStateArray.length; i++) {
+    cellStateArray[i] = i % 2;
+}
+device.queue.writeBuffer(cellStateStorage[1], 0, cellStateArray);
+
+const RULE = parseRuleString(EventManager.ruleString);
+const { ruleArray, ruleStorage } = rules(RULE);
+console.log(ruleArray)
+device.queue.writeBuffer(ruleStorage, 0, ruleArray);
+
+// FORMAT CANVAS
+const context = canvas.getContext("webgpu");
+const canvasFormat = navigator.gpu.getPreferredCanvasFormat();
+context.configure({
+    device: device,
+    format: canvasFormat,
+});
+
+// VERTEX SETUP for a square
+const vertices = new Float32Array([
     // X,    Y,
     -0.8, -0.8, // Triangle 1
     -0.8, 0.8,
@@ -51,36 +101,24 @@ const SQUARE_VERTICIES = new Float32Array([
     -0.8, -0.8,
 ]);
 
-
-let step = 0; // How many compute passes have been made
-
-// DEVICE SETUP - ran into issues making it a func
-// CONFIGURE CANVAS
-const context = canvas.getContext("webgpu");
-const canvasFormat = navigator.gpu.getPreferredCanvasFormat();
-context.configure({
-    device: device,
-    format: canvasFormat,
+const vertexBuffer = device.createBuffer({
+    label: "Cell vertices", // Error message label
+    size: vertices.byteLength,
+    usage: GPUBufferUsage.VERTEX | GPUBufferUsage.COPY_DST,
 });
 
+device.queue.writeBuffer(vertexBuffer, /*bufferOffset=*/0, vertices);
 
-// DRAWING STUFF setup code 
-// vertex setup for a square
-const {vertexBuffer, vertexBufferLayout} = BufferManager.loadShapeVertexBuffer(device, SQUARE_VERTICIES);
+const vertexBufferLayout = {
+    arrayStride: 8, // 32bit = 4 bytes, 4x2 = 8 bytes to skip to find next vertex
+    attributes: [{
+        format: "float32x2", // two 32 bit floats per vertex
+        offset: 0,
+        shaderLocation: 0, // Position, see vertex shader
+    }]
+}
 
-// load shader code for drawing operations
-const cellShaderModule = device.createShaderModule({
-    label: 'shader that draws',
-    code: guiShader
-});
-
-
-// COMPUTE SHADER setup code
-// 1. load grid data into buffers etc. relevant to the simulation inc. parse rulestring 
-// 2. define the layout of loaded binary data 
-const bindGroupLayout = BufferManager.createBindGroupLayout(device);
-
-// load shader module for running simulation
+// COMPUTE SHADER MODULE
 const simulationShaderModule = device.createShaderModule({
     label: 'shader that computes next state',
     code: computeShader,
@@ -88,14 +126,61 @@ const simulationShaderModule = device.createShaderModule({
 }
 );
 
+// CELL SHADER MODULE
+const cellShaderModule = device.createShaderModule({
+    label: 'shader that draws',
+    code: guiShader
+});
+
+// COMPUTE SHADER RESOURCE BINDING LAYOUT
+const bindGroupLayout = device.createBindGroupLayout({
+    label: "Cell Bind Group Layout",
+    entries: [
+        {
+            binding: 0,
+            visibility: GPUShaderStage.VERTEX | GPUShaderStage.COMPUTE | GPUShaderStage.FRAGMENT,
+            buffer: {} // Grid uniform buffer
+        },
+
+        {
+            binding: 1,
+            visibility: GPUShaderStage.VERTEX | GPUShaderStage.COMPUTE | GPUShaderStage.FRAGMENT,
+            buffer: { type: "read-only-storage" } // Cell state input buffer
+        },
+
+        {
+            binding: 2,
+            visibility: GPUShaderStage.COMPUTE,
+            buffer: { type: "storage" } // Cell state output buffer
+        },
+
+        {
+            binding: 3,
+            visibility: GPUShaderStage.COMPUTE,
+            buffer: { type: "read-only-storage" } // Ruleset
+        }
+    ]
+});
+
+
 // PIPE LAYOUT
 const pipelineLayout = device.createPipelineLayout({
     label: "Cell Pipeline Layout",
     bindGroupLayouts: [bindGroupLayout],
 });
 
+// SIMULATION PIPELINE
+const simulationPipeline = device.createComputePipeline({
+    label: "Simulation pipeline",
+    layout: pipelineLayout,
+    compute: {
+        module: simulationShaderModule,
+        entryPoint: "computeMain",
+    }
+});
+
 //RENDER PIPELINE            
-const renderPipeline = device.createRenderPipeline({
+const cellPipeline = device.createRenderPipeline({
     label: "Cell pipeline",
     layout: pipelineLayout,
     vertex: {
@@ -112,29 +197,100 @@ const renderPipeline = device.createRenderPipeline({
     }
 });
 
-// SIMULATION PIPELINE
-const simulationPipeline = device.createComputePipeline({
-    label: "Simulation pipeline",
-    layout: pipelineLayout,
-    compute: {
-        module: simulationShaderModule,
-        entryPoint: "computeMain",
-    }
+// setup bind groups
+let bindGroups = [
+    createBindGroup("Cell renderer bind group A", uniformBuffer, cellStateStorage[0], cellStateStorage[1], ruleStorage),
+    createBindGroup("Cell render bind group B", uniformBuffer, cellStateStorage[1], cellStateStorage[0], ruleStorage)
+];
+
+
+// INITIAL CANVAS SETUP
+const encoder = device.createCommandEncoder();
+const pass = encoder.beginRenderPass({
+    colorAttachments:
+        [{
+            view: context.getCurrentTexture().createView(),
+            loadOp: "clear",
+            clearValue: { r: 0, g: 0, b: 0, a: 1 }, // New line
+            storeOp: "store",
+        }]
 });
 
-const { bindGroups, uniformBuffer, cellStateStorage, ruleStorage } = BufferManager.initialiseComputeBindgroups(device, renderPipeline, GRID_SIZE, INITIAL_STATE, parseRuleString(EventManager.ruleString) );
-
-
-// INITIAL CANVAS SETUP, 1st render pass
-const encoder = device.createCommandEncoder();
-renderPass(encoder);
+// Draw the features
+pass.setPipeline(cellPipeline);
+pass.setVertexBuffer(0, vertexBuffer);
+pass.setBindGroup(0, bindGroups[step % 2]);
+pass.draw(vertices.length / 2, GRID_SIZE * GRID_SIZE); // 6 vertices, 12 floats
+pass.end();
 
 // Finish the command buffer and immediately submit it.
 device.queue.submit([encoder.finish()]);
 
-EventManager.bindEvents();
+// get and bind events from html
+document.getElementById('play').addEventListener('click', EventManager.playPause);  // play pause button
+document.getElementById('next').addEventListener('click', EventManager.moveOneFrame); // move one frame button
+document.getElementsByTagName("body")[0].addEventListener("keydown", EventManager.keyListener); // key presses
+document.getElementById('submitInput').addEventListener('click', EventManager.updateRuleString); // new rule string input button
+document.getElementById('speedInput').addEventListener('click', () => {
+    EventManager.updateSpeed();
+    clearInterval(Interval);
+    console.log(EventManager.updateInterval)
+    Interval = setInterval(updateLoop, EventManager.updateInterval)
+}
 
-EventManager.updateLoop = () => {
+); // change speed
+
+// iterative update for cells
+console.log(EventManager.updateInterval);
+var Interval = setInterval(updateLoop, EventManager.updateInterval);
+EventManager.forcedUpdate = updateLoop;
+
+function createBindGroup(label, uniformBuffer, cellStateA, cellStateB, ruleStorage) {
+    return device.createBindGroup({
+        label: label,
+        layout: cellPipeline.getBindGroupLayout(0),
+        entries: [
+            { binding: 0, resource: { buffer: uniformBuffer } },
+            { binding: 1, resource: { buffer: cellStateA } },
+            { binding: 2, resource: { buffer: cellStateB } },
+            { binding: 3, resource: { buffer: ruleStorage } },
+        ],
+    });
+}
+
+function parseRuleString(ruleString) {
+    // ruleString is given by the user. it is a string
+    let RULE = new Uint32Array(1)
+    let slashFlag = false // tells us whether we are before or after the / symbol
+    for (let i = 0; i < ruleString.length; i++) {
+        let char = ruleString[i];
+        if (char === "/") {
+            slashFlag = !slashFlag
+            continue
+        }
+        let num = Number(char) // the character is indeed a number
+        switch (slashFlag) {
+            case false: // before "/" sign. survival case
+                RULE[0] += 2 ** (num + 9)
+                break;
+            case true: // after "/" sign. birth case
+                RULE[0] += 2 ** num
+        }
+    }
+    return RULE
+}
+
+function rules(rule) {
+    const ruleArray = rule;
+    const ruleStorage = device.createBuffer({
+        label: "Rule Storage",
+        size: ruleArray.byteLength,
+        usage: GPUBufferUsage.STORAGE | GPUBufferUsage.COPY_DST,
+    });
+    return { ruleArray, ruleStorage };
+}
+
+function updateLoop() {
 
     // Only permitted to run if one frame is wanted or
     if (!EventManager.oneFrame) {
@@ -147,60 +303,21 @@ EventManager.updateLoop = () => {
 
     // check for new rule string
     if (EventManager.newRuleString) {
-        const ruleStorage = BufferManager.setRuleBuffer(device, parseRuleString(EventManager.ruleString));
-        bindGroups[0] = BufferManager.createBindGroup(device, renderPipeline, "Cell renderer bind group A", uniformBuffer, cellStateStorage[0], cellStateStorage[1], ruleStorage);
-        bindGroups[1] = BufferManager.createBindGroup(device, renderPipeline, "Cell render bind group B", uniformBuffer, cellStateStorage[1], cellStateStorage[0], ruleStorage);
-  
+        const { ruleArray, ruleStorage } = rules(parseRuleString(EventManager.ruleString))
+        console.log(ruleArray)
+        device.queue.writeBuffer(ruleStorage, 0, ruleArray);
+        bindGroups = [
+            createBindGroup("Cell renderer bind group A", uniformBuffer, cellStateStorage[0], cellStateStorage[1], ruleStorage),
+            createBindGroup("Cell render bind group B", uniformBuffer, cellStateStorage[1], cellStateStorage[0], ruleStorage)
+        ];
         EventManager.newRuleString = false // toggle off
 
     }
 
+
     const encoder = device.createCommandEncoder();
 
     // CREATE COMPUTE TOOL & PERFORM COMPUTATION TASKS
-    computePass(encoder);
-
-    // CREATE DRAW TOOL & SET DEFAULT COLOR (BACKGROUND COLOR)
-    step++;
-    renderPass(encoder);
-
-    // Finish the command buffer and immediately submit it.
-    device.queue.submit([encoder.finish()]);
-}
-
-
-
-// start iterative update for cells
-EventManager.currentTimer = setInterval(EventManager.updateLoop, EventManager.updateInterval); // Interval is accessed from an externally called function
-EventManager.forcedUpdate = EventManager.updateLoop;
-
-
-
-
-// FUNCTIONS for convenient break-up of code
-    // Can't be removed because relies on a ton of 
-    // definitions from this chunk of code.
-function renderPass(encoder){
-    const renderPass = encoder.beginRenderPass({
-        colorAttachments:
-            [{
-                view: context.getCurrentTexture().createView(),
-                loadOp: "clear",
-                clearValue: { r: 0, g: 0, b: 0, a: 1 }, // New line
-                storeOp: "store",
-            }]
-    });
-    
-    // Draw the features
-    renderPass.setPipeline(renderPipeline);
-    renderPass.setVertexBuffer(0, vertexBuffer);
-    renderPass.setBindGroup(0, bindGroups[step % 2]);
-    renderPass.draw(SQUARE_VERTICIES.length / 2, GRID_SIZE * GRID_SIZE); // 6 vertices, 12 floats
-    renderPass.end();
-}
-
-
-function computePass(encoder){
     const computePass = encoder.beginComputePass();
 
     computePass.setPipeline(simulationPipeline);
@@ -210,4 +327,26 @@ function computePass(encoder){
     computePass.dispatchWorkgroups(workgroupCount, workgroupCount);
 
     computePass.end();
+
+    // CREATE DRAW TOOL & SET DEFAULT COLOR (BACKGROUND COLOR)
+    step++;
+    const pass = encoder.beginRenderPass({
+        colorAttachments:
+            [{
+                view: context.getCurrentTexture().createView(),
+                loadOp: "clear",
+                clearValue: { r: 0, g: 0, b: 0, a: 1 }, // New line
+                storeOp: "store",
+            }]
+    });
+
+    // DRAW THE FEATURES
+    pass.setPipeline(cellPipeline);
+    pass.setVertexBuffer(0, vertexBuffer);
+    pass.setBindGroup(0, bindGroups[step % 2]);
+    pass.draw(vertices.length / 2, GRID_SIZE * GRID_SIZE); // 6 vertices, 12 floats
+    pass.end();
+
+    // Finish the command buffer and immediately submit it.
+    device.queue.submit([encoder.finish()]);
 }
