+++ CAs/LifeLike/managers/DeviceManager.js
