+++ CAs/LifeLike/managers/PipelineManager.js
