+++ CAs/LifeLike/managers/ShaderManager.js
